// app/src/generated/models/Ben_ltkvdtscenariosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkvdtscenariosstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkvdtscenariosstatecode = keyof typeof Ben_ltkvdtscenariosstatecode;
export const Ben_ltkvdtscenariosstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkvdtscenariosstatuscode = keyof typeof Ben_ltkvdtscenariosstatuscode;

export interface Ben_ltkvdtscenariosBase {
  ben_assumedjson?: string;
  ben_at?: string;
  ben_author?: string;
  ben_authorid?: string;
  ben_ltkvdtscenarioid: string;
  ben_name?: string;
  ben_period?: string;
  ben_scenarioid: string;
  ben_site?: string;
  ben_togglesjson?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkvdtscenariosstatecode;
  statuscode?: Ben_ltkvdtscenariosstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkvdtscenarios extends Ben_ltkvdtscenariosBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}